BUG Page not loading
when trying to find module "Grid", the app does not respond
